Seeder

+++ db/migrate/20180529200416_create_nodes_table.rb
@@ -2,9 +2,9 @@ class CreateNodesTable < ActiveRecord::Migration[5.1]
   def change
     create_table :nodes do |t|
       t.integer('trail_id')
-      t.integer('latitude')
-      t.integer('longitude')
-      t.integer('elevation')
+      t.float('latitude')
+      t.float('longitude')
+      t.float('elevation')
     end
   end
 end
